+++ unstructured/__version__.py
@@ -1,1 +1,1 @@
-__version__ = "0.11.4-dev13"  # pragma: no cover
+__version__ = "0.11.4-dev14"  # pragma: no cover
