+++ unstructured/partition/doc.py
@@ -2,7 +2,7 @@ import os
 import tempfile
 from typing import IO, Any, List, Optional
 
-from unstructured.chunking.title import add_chunking_strategy
+from unstructured.chunking import add_chunking_strategy
 from unstructured.documents.elements import Element, process_metadata
 from unstructured.file_utils.filetype import FileType, add_metadata_with_filetype
 from unstructured.partition.common import (
